Adds a check to SetAntennaPatternVal() to make sure the pattern array has been initialised.
Now that the initialisation of the pattern array is called up when
the antenna input file is read, it's possible that the SetAntennaVal()
function can bee called before the array has been allocated.  This commit
add a fix to check if the array is NULL and if so, allocates a data structure
array for a single frequency.