Removes the '.' character on the RPT file prefix for Linux.
The '.' character at the beginning of the auto generated file name
creates a hidden file on Linux.  When run as a conventional applictaion
it's just a hidden file.  When run as a snap it casuses and error
as snaps are not allowed access to hidden files in the users home area.