Fix sscanf usage
Compilers (at least gcc and clang) report type errors when converting
using %s and %[...] to pointers to char*. Presumably these are typos
with the intent being to convert into the buffers.

This commit changes those, and also changes the instances where sscanf
is scanning from instr into instr, which might be OK but seems
confusing (and is likely undefined begaviour technically). These are
changed to use two buffers which are used alternately. (Again, I think
that matches the intent though I'm not entirely confident.)