fix common.h vs. Common.h
P372 has both common.h and Common.h, with identical contents. Some files
include one, some include the other. Besides being untidy, this actually
causes a build failure if building for Linux in Mac or Windows Docker,
because only one copy gets checked out onto the host filesystem, and the
other variant is "not found" when the compiler goes looking for it.

Initial-uppercase names seem to be the convention here, so remove
common.h and correct references to point to Common.h.